Switched location of OnPostDeleteLibrary
